Add color scope.
Same device as Spacewar consoles.  See ITS hardware memo 6 for details.

CONO 420: Bits 404040 enable three different sets of switches.  400 is
Spacewar consoles and switches in the color scope room.  The other two
are not supported yet.  Bits 202020 enable setting red, green, and
blue.  Bits 171717 are 4 bits each for the colors.

DATAO 420: The lowest 9 bits are Y, and the next 9 bits are X.

The wire protocol used to send data from the pdp6 simulator transmits
32 bits at a time, in little endian format.  The lowest 10 bits are X,
with the top bit always zero.  The next 10 bits are Y.  The top 12
bits are four each for R, G, and B.